transformers/face_detection/requirements.txt
fastapi
uvicorn[standard]
gunicorn
aiohttp
numpy
opencv-python
aiofiles
kaggle==1.5.16
webdataset==0.2.86
Pillow==10.0.0